v1.0.0
New temporary UI, some API changes

+++ app/src/main/java/com/redc4ke/taniechlanie/data/AlcoObject.kt
@@ -1,15 +1,16 @@
 package com.redc4ke.taniechlanie.data
 
 import java.io.Serializable
+import java.math.BigDecimal
 
 data class AlcoObject(
         var id: Int,
         var name: String,
-        var minPrice: Float,
-        var maxPrice: Float?,
-        var promoPrice: Float?,
+        var minPrice: BigDecimal,
+        var maxPrice: BigDecimal?,
+        var promoPrice: BigDecimal?,
         var volume: Int,
-        var voltage: Float,
+        var voltage: BigDecimal,
         var shop: ArrayList<Int>,
         var categories: ArrayList<Int>,
         var photo: String?
